# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flight Advisor Api Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "jstojiljkovic941@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/flight-advisor/v1/flights/cheapest": {
            "get": {
                "tags": [
                    "Flights"
                ],
                "summary": "Returns the cheapest  route from city A to B",
                "description": "",
                "operationId": "cheapest",
                "parameters": [
                    {
                        "name": "city1",
                        "in": "query",
                        "description": "City you are flying from",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "city2",
                        "in": "query",
                        "description": "City you are flying to",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Finds cheapest flight between two cities",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "start": {
                                                        "description": "Finds cheapest flight between two cities",
                                                        "properties": {
                                                            "airline": {
                                                                "description": "Finds cheapest flight between two cities",
                                                                "type": "string",
                                                                "example": "Hartsfield Jackson Atlanta International Airport"
                                                            },
                                                            "city": {
                                                                "description": "Finds cheapest flight between two cities",
                                                                "type": "string",
                                                                "example": "Atlanta"
                                                            },
                                                            "country": {
                                                                "description": "Finds cheapest flight between two cities",
                                                                "type": "string",
                                                                "example": "United States"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "end": {
                                                        "description": "Finds cheapest flight between two cities",
                                                        "properties": {
                                                            "name": {
                                                                "description": "Finds cheapest flight between two cities",
                                                                "type": "string",
                                                                "example": "Dothan Regional Airport"
                                                            },
                                                            "city": {
                                                                "description": "Dothan",
                                                                "type": "string"
                                                            },
                                                            "country": {
                                                                "description": "Finds cheapest flight between two cities",
                                                                "type": "string",
                                                                "example": "United States"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "price": {
                                                        "description": "Finds cheapest flight between two cities",
                                                        "type": "float",
                                                        "example": "42.04"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Returns when validation fails",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Finds cheapest flight between two cities",
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Finds cheapest flight between two cities",
                                            "type": "string",
                                            "example": "Invalid credentials."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Returns when cities with the provided id does not exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Finds cheapest flight between two cities",
                                            "type": "string",
                                            "example": "Airport in the city Nis does not exist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basic": []
                    }
                ]
            }
        },
        "/flight-advisor/v1/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Registers a new user",
                "description": "",
                "operationId": "store",
                "requestBody": {
                    "description": "Add user object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "john_doe"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "first_name": {
                                            "type": "string",
                                            "example": "John"
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": "Doe"
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "john_doe"
                                        },
                                        "id": {
                                            "type": "int",
                                            "example": "1"
                                        },
                                        "created_at": {
                                            "description": "Initial creation timestamp",
                                            "type": "string",
                                            "format": "date-time",
                                            "readOnly": "true"
                                        },
                                        "updated_at": {
                                            "description": "Last update timestamp",
                                            "type": "string",
                                            "format": "date-time",
                                            "readOnly": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Returns when username is already taken or validation fails",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}